Factor out common lesson plan translations
