+++ client/app/bundles/course/lesson-plan/pages/LessonPlanEdit/index.jsx
@@ -1,36 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { defineMessages, FormattedMessage } from 'react-intl';
+import { FormattedMessage } from 'react-intl';
+import translations from 'course/lesson-plan/translations';
 import MilestoneRow from './MilestoneRow';
 import ItemRow from './ItemRow';
-
-const translations = defineMessages({
-  type: {
-    id: 'course.lessonPlan.LessonPlanEdit.type',
-    defaultMessage: 'Type',
-  },
-  title: {
-    id: 'course.lessonPlan.LessonPlanEdit.title',
-    defaultMessage: 'Title',
-  },
-  startTime: {
-    id: 'course.lessonPlan.LessonPlanEdit.startTime',
-    defaultMessage: 'Start Time',
-  },
-  bonusEndTime: {
-    id: 'course.lessonPlan.LessonPlanEdit.bonusEndTime',
-    defaultMessage: 'Bonus End Time',
-  },
-  endTime: {
-    id: 'course.lessonPlan.LessonPlanEdit.endTime',
-    defaultMessage: 'End Time',
-  },
-  published: {
-    id: 'course.lessonPlan.LessonPlanEdit.published',
-    defaultMessage: 'Published',
-  },
-});
 
 class LessonPlanEdit extends React.Component {
   static propTypes = {
@@ -48,9 +22,9 @@ class LessonPlanEdit extends React.Component {
         <tr>
           {rowFor('type')}
           {rowFor('title')}
-          {rowFor('startTime')}
-          {rowFor('bonusEndTime')}
-          {rowFor('endTime')}
+          {rowFor('startAt')}
+          {rowFor('bonusEndAt')}
+          {rowFor('endAt')}
           {rowFor('published')}
         </tr>
       </thead>
