+++ client/app/bundles/course/lesson-plan/translations.js
@@ -1,0 +1,30 @@
+import { defineMessages } from 'react-intl';
+
+const translations = defineMessages({
+  type: {
+    id: 'course.lessonPlan.type',
+    defaultMessage: 'Type',
+  },
+  title: {
+    id: 'course.lessonPlan.title',
+    defaultMessage: 'Title',
+  },
+  startAt: {
+    id: 'course.lessonPlan.startAt',
+    defaultMessage: 'Start At',
+  },
+  bonusEndAt: {
+    id: 'course.lessonPlan.bonusEndAt',
+    defaultMessage: 'Bonus End At',
+  },
+  endAt: {
+    id: 'course.lessonPlan.endAt',
+    defaultMessage: 'End At',
+  },
+  published: {
+    id: 'course.lessonPlan.published',
+    defaultMessage: 'Published',
+  },
+});
+
+export default translations;
